Added module for OpenLink Virtuoso test container
The test container runs the tenforce/virtuoso image. The extended
database container class provides both jdbc and the sparql service urls.
In addition the container database driver now uses a more generic
pattern that is not just limited to the provided modules. Thus it is
easier to add own modules without modifying the regex pattern.