Added MinGW compatible paths and choose paths based on OS
The prior class PathOperations is renamed to PathUtils that now supports
to convert a windows path into a MinGW compatible path. Furthermore,
only some paths need to be converted. Thus a new method
"addHostSystemAwareFileSystemBind" for GenericContainer is introduced
that only transforms paths if run on a Windows system.

Fixes issue #84 .