Added switches for OS specific tests
Tests using OS dependent features like right management are not executed
on an OS not supporting these features. In this case the ini-file for
MySQL databases that is dynamically loaded into the container needs to
have unix like limited rights which are not available on Windows, thus
the ini file is not loaded and the tests fail.