closes #182 set mode on tarEntry from transferable
Restructure control flow in container tryStart so that logs are output in both startup failure and startup timeout scenarios

Add container logging from the point of startup (not used for debugging in this instance, but useful feature)

Ensure that filePermissions test container is treated start-stop rather than using the default startup check, which expects the container to start and run continuously.

Capture and wait on logs from the point of container start to allow tests to assert output.

SQUASHME

Refactor container startup failure detection for clarity

Add withLogConsumer method to Container and GenericContainer with basic javadocs (improves but does not resolve #113)