More accurate log processing. (#643)
* More accurate log processing.

The Docker logs can come in non normalized form. The line breaks can be not at end of line, the frame's boundary can split multibyte unicode symbols in the middle and so on.

This patch address this problems and normalize logs so, that they can be processed strictly line by line.

To achive this, the extra base class BaseConsumer have been added. That class normalize the incoming in method accept() logs and forward it to method process() for it's child classes. Other *Consumer classes have been reworked to be child to the BaseConsumer class and only do own work.

Adititionally, BaseConsumer have new withRemoveAnsiCodes(boolean) method for ability to disable ANSI color codes removing (true by default).

* Move normalization logic directly to FrameConsumerResultCallback

It's help to not ruin the API of *Consumer classes.

* Added some tests for log normalization in class FrameConsumerResultCallback.

* Fix Codacity warning

* More test for the FrameConsumerResultCallback class

* Stabilize tests

* Log consumers, that not derive BaseConsumer class, is not receive color codes now. Added record to the changelog.

* One more test for FrameConsumerResultCallback class.

* Fixes due to the code review recommendations

* Fixes due to the code review recommendations (Part 2)

* One more use case for FrameConsumerResultCallback class. And unit test for it.
If StreamType is STDERR or STDOUT the log always have newline at line end. Therefore preprocessor should trim it to be consistent with RAW type processing.

* Roll back previouse change doe to failing tests. Doing newline trimming directly in Slf4jLogConsumer class.

* Fixes due to the code review recommendations (Part 3)