Improve fluent builders
Currently many containers (JDBC based, Nginx, Vault) are not very great
as fluent builders. They are all parametrized, that means the user
has to instance them with the sqare brackets to denote the type.
That's not intuitive, and the Testcontainers docs don't mention it
anywhere. Moreover, the compiler still complains about raw types,
because we can't declare a container variable parametrized by itself.

If we don't parametrize the container, than we can't invoke the methods
of `GenericContainer` without losing the call chain. The compiler
doesn't know about the concrete type of the builder. For example,
for `MySQLContainer` this causes a compiler error:

```
MySQLContainer container = new MySQLContainer("mysql:5.7")
.withClasspathResourceMapping("/mysql-initdb", "/docker-entrypoint-initdb.d", BindMode.READ_ONLY)
.withDatabaseName("test-warehouse");
```

A solution is not to parametrize concrete containers and let the type
propagate to `GenericContainer`.