Allow testcontainer to wait for a specific port (#703)
* Allow testcontainer to wait for a specific port

*Motivation*

A container might expose multiple ports for different purposes. for example, it can have a port for http endpoint,
while having another port for grpc endpoint. So when doing a liveness check, it would be better to allow configuring
checking a specific port, rather than checking the first exposed port.

*Modification*

Modify the `HttpWaitStrategy` to allow configuring the liveness port.

* - change logging from info to trace
- add an item in CHANGELOG.md

* Update CHANGELOG.md