Remove the need for explicit casting for MySQLContainer
At $WORK we use TestContainer, and it's very annoying to do the
`MySQLContainer mysql = (MySQLContainer) new MySQLContainer()` dance every
time we want to create a MySQL container. Plus, the compiler complains about
raw use of a parametrized class cast, so we have to add `@SupressWarning("rawtypes")`,
which adds another level of verbosity. It would be nice if we could create
a MySQL container just by `MySQLContainer mysql = new MySQLContainer()`.