Allow `HostnameVerifier` to be overridden
Modifies `HttpWaitStrategy` to allow `HostnameVerifier` to be overridden
for SSL connections.

This can be useful, for instance, when the Subject Alternative Name in
the contained service's SSL certificate differs from the hostname that
`HttpWaitStrategy` is using to make requests to the service. In this
case, a `NoopHostnameVerifier` allows the `HttpWaitStrategy` to execute
successfully despite the inconsistency between the service's SSL
certificate and the container's hostname (assuming one is ok with the
related security concerns of not verifying hostnames).