added read timeout to http wait strategy (#2058)
This commit adds the possibility to add a read timeout to the HTTP call
for the HTTP wait strategy.
This is necessary, because with the default Docker settings for Mac
can lead to a blocking call to `HttpURLConnection.connect()`. The follow
up issue is, that the container will never get healthy and due to
automated integration tests the tests are not able to run.
The main cause of the problem is the lazy connection establishment
during setting up the container. So the time between starting the
container and starting the wait strategy checks.

The read timeout lets the HTTP call fail, so the `retryUntilSuccess()`
is able to send another HTTP call in order to reestablish the connection
and not block the whole thread.