Changed Class.getClass.getClassLoader() and object.getClass().getClassLoader() to use Thread.currentThread().getContextClassLoader() instead. The reason for this is that it allows the use of parameterized tests where, for instance, we want to check the behavior of our code when using multiple versions of the JDBC driver.
That's not possible with the current approach since it will only
allow the use of a driver in the classpath. With the change the
developer can create a new URLClassLoader that points to the specific
JDBC driver and set it with Thread.currentThread().setContextClassLoader().

An alternative is modifying the JDBC container so it will accept a Driver
instance. This could be done in addition to this change.