Generate filesystem friendly name from display name (#1326)
Generating a filesystem friendly name in a Junit Jupiter test is a bit tricky.
Since tests can be generated dynamically class and/or test method may not be available.
The display name provided by the ExtensionContext on the other hand may use characters
that are not filesystem safe.
This approach removes all characters from the display name that are not in a restricted
set of allowed characters. However this may lead to name clashes if two tests have a
display name that only differs in characters that are removed from the display name.