Support TestLifecycleAware-ness of containers started by the JUnit Jupiter integration (#1326)
* Update BrowserWebDriverContainer to honor existing no_proxy setting

* Update test to only start one container per test

* Use constant for no_proxy key

* Cleanup test implementation (#929)

* Add signalling of TestLifecycleAware containers.

Allow containers like WebBrowserContainers to initialize and/or finalize
before/after tests.

* #1326 Add early draft to test signalling of TestLifecycleAware containers

* Add test for post condition when signalling lifecycleaware containers

This kind of test is a bit tricky since the post condition occurs
after the original test has been finished. Also it's not nice to
pass data between two tests.

* Update test for lifecycle aware containers to cover shared case (#1326)

In order to check that the afterAll callback has signalled
lifecycleaware containers correctly a second extension is used. The
order of the extension annotation ensures that the assertion is run
after the extension under test.

* Update test for lifecycle aware containers to cover shared case (#1326)

To test the beforeAll() case the assertion has to be called from
within the test class since it's called after all extensions.

* Fix formatting (#1326)

* Use lighter container for testing (#1326)

* Separate store and collect of shared lifecycle-aware-containers (#1326)

* Add tests for ordering and capturing test exceptions (#1326)

* Make lifecycle tests independent of timing (#1326)

Calls to the lifecycle methods are now recorded in an ordered
list that is then used to test the correct number and order
of calls. This makes the test independent of timing.
Unfortunately it's still required to execute tests in
a deterministic order.

For a better separation of test concerns
tests for the lifecycle methods and exception
capturing have been moved into separate
test classes.

* Make mock now implements Startable (#1326)

There is no need to start a container since only
the TestLifecycleAware is important.

* Add AssertJ dependency (#1326)

We want to use AssertJ for some tests.

* Migrate assertions of TestLifecycleAwareMethodTest to AssertJ (#1326)

* Update generation of filesystem friendly description (#1326)

* Separated tests for filesystem friendly filename (#1326)

* Use lombok to improve readability (#1326)

* Generate filesystem friendly name from display name (#1326)

Generating a filesystem friendly name in a Junit Jupiter test is a bit tricky.
Since tests can be generated dynamically class and/or test method may not be available.
The display name provided by the ExtensionContext on the other hand may use characters
that are not filesystem safe.
This approach removes all characters from the display name that are not in a restricted
set of allowed characters. However this may lead to name clashes if two tests have a
display name that only differs in characters that are removed from the display name.

* Generate filesystem friendly name from URLEncoded unique id (#1326)

Co-authored-by: Richard North <rich.north@gmail.com>
Co-authored-by: Kevin Wittek <kiview@users.noreply.github.com>