Rewrite Couchbase module. closes #2447 (#2491)
* Rewrite couchbase module. closes #2447

This changeset completely reworks the couchbase module and hopefully
greatly improve the out-of-the-box experience. Note that this is a
breaking change over the previous code because by intention it does
NOT depend on SDK 2 so you can test SDK 2 and 3 with it at the same
time.

Highlights:

 - Removed the need for a SDK, so both 2 and 3 can be used with it.
 - Updated to 6.5.0 baseline and using alternate addresses for
   "proper" port exposure without having to rely on the socat
   proxy container like the previous version had to.
 - Allows to define which services should be exposed and handles
   states automatically (i.e. will not try to create the primary
   index if the query service is not enabled).

Note that a bunch of tests have been removed since they are not
adequate anymore. A side effect of the alternate address change
is that older servers cannot be used. 6.5.0 is available in both
CE and EE, and Couchbase in general allows EE versions to be used
in development and testing so we should use it if we can.

* Wait until query respsonds with 200

* fixes

* add `getBootstrap*DirectPort()` methods, review fixes

* Restore `getConnectionString()`

* Apply suggestions from code review

Co-Authored-By: Kevin Wittek <kiview@users.noreply.github.com>

* Update docs

* Update docs (2)

Co-authored-by: Michael Nitschinger <michael@nitschinger.at>
Co-authored-by: Kevin Wittek <kiview@users.noreply.github.com>