Deprecate ambiguous constructors (#2839)
Our constructors for `GenericContainer` and all the classes that inherit from it tend to be a bit inconsistent. In some places, a no-args constructor is favoured, in other places a String parameter is favoured - which can be either a full image name or a tag.

This raises confusion, and in the case of an image tag being the parameter, it restricts which images people can easily use.

Another problem this raises is when the container class specifies a default tag version: this frequently falls out of date. We're stuck for options: some users (and image maintainers) would like to see newer image versions being used, but this could be a breaking change for existing users. This is especially risky when a major version bump is proposed.

This change aims to:

- Deprecate 'ambiguous' constructors for containers - any no-args or String-arg constructor gets this treatment.
- Introduce a new constructor which accepts a `DockerImageName` object. This class is preferred to a String because its meaning is unambiguous.
- Enhance `DockerImageName` with a static factory method and a convenience `withTag` method so that users can easily create and derive new instances.
- Update all modules to follow suit