Remove use of implicit network for multi-container Kafka
The use of implicit networks in the Kafka container (like other
containers) is no longer correct. In order for multiple containers to be
able to communicate with one another via an internal Docker network, it
is necessary for all of them to be given an explicit reference to a
Network object.