Jqwik Support: Shared and Restart containers
The TestContainersExtension of Jupiter was adapted to support the BeforeContainer, AfterContainer and AroundProperty hooks. They work in the same way as the Jupiter pendants. However, the Container annotation was renamed to TestContainer because the name container is already used in Jqwik for any class which contains properties.

Further, the store in Jqwik is not bound to its current context and not a key-value store. Though, it can be scoped to the overall run, property or the current try.