couchbase: add support for loading sample buckets
Couchbase provides 3 different sample buckets which are usually
enabled manually in the UI for experimentation. It is also possible
to load them via the REST API - which is what this changeset
implements.

Loading a sample bucket allows new users to toy around with the
SDK APIs quickly without having to come up with sample data
and indexes themselves.

Note that the "wait until the bucket is ready" logich as been
refactored int its own method so it can be used by both the
regular buckets as well as the sample bucket logic.

Also note that additional logic has been put into place for
the sample buckets to get the document count (which is
fixed) so that we know the bucket is actually fully
populated before returning back to the user.