Ability to specify the image pull policy in the DockerComposeContainer (#4601)
The "withImagePullPolicy" method has been added to the DockerComposeContainer.
This image pull policy will only be used if the value of the property DockerComposeContainer#pull is true.
This setting only determines the behavior of DockerComposeContainer and does not affect the image pulling process that docker-compose can execute before starting containers.