[couchbase] Ignore transient primary index creation error.
This changeset makes sure that when a primary index is created, a transient
error is ignored since the indexing engine will automatically retry the
statement later.

Transient errors can happen because we just created the bucket and the
indexer might not have caught up to that bucket just yet, but it will
eventually.

In addition, a check has been added afterwards to make sure that the
primary index is online (so, ready for the user to query) before
returning to the caller, improving the overall stability of queries.