Update gitlab_ci.md
I had problems to get my configuration working for testcontainers in combination with gitlab. On my machine everything did run completely fine. But in Gitlab CI I got problems with Ryuk and also my mongo container couldn't be found. This was due to a wrong resolved ip address. 
I found out when setting  the `TESTCONTAINERS_HOST_OVERRIDE` to `host.docker.internal` that it helped me a lot to solve these issues (since testcontainers exposes the ports on the host). 

Since my configuration is quite generic, I thought it would be nice to include it in the docs so other people could also benefit.