[couchbase] Explicitly configure service quotas (and allow customization).
This changeset explicitly configures the service memory quotas to their
documented limits by default. This ensures that if a service is enabled,
it does not consume more memory than absolutely needed. This is important
because some services (like analytics) try to grab memory available which
is good for production but of limited use in a testcontainers setting.

A new API is also added which allows to customize the memory quota for
each service should the test case require more than the minimum (which
for most services is 256MB anyways).