[couchbase] Add support for ephemeral buckets.
Ephemeral buckets, other than the default Couchbase buckets are
purely in-memory and as a result have different tradeoffs in a
production setup. For testing though, it can make sense to use
them instead of the default ones if no views are needed and
less disk access in a CI environment is needed (which may
speed up tests).