Pull k3s module
due to previously unnoticed issues with pod startup in the latest Docker Desktop for Mac.

We'll need to investigate and fix.

Likely cgroupv2 related; the workaround we had seems to allow the API server to start up properly but does not allow pod container sandboxes to be created. Tests obviously need to be extended to cover pod creation.