Elasticsearch: Don't throw exception on missing CA cert file
The current check throws an exception if the version is supposed to be
greater than 8 and the ca cert file is missing. This posed two problems:

1. A custom docker image might have a 'latest' version, and thus is
   newer than version 8
2. A user may have configured their own CA in a different path

In order to accomodate for this, a missing copy of a the ca crt file
will log a warning, but not throw an exception and leave the container
unstarted. There is also another public method, allowing to read the
cert, called `extractCert`.

Closes #5264