[LocalStack] Deprecate methods using aws sdk v1 (#5489)
Currently, `LocalStackContainer` depends on aws sdk v1. We have received
feedback about moving to aws sdk v2. However, we think that the
class already provides the enough information to be used such as
`getEndpointOverride()`, `getRegion()`, `getAccessKey()` and `getSecretKey`.
So, we are deprecating `getDefaultCredentialsProvider()` and `getEndpointConfiguration`
which are going to be removed in the feature and `LocalStack` module will
not depend on any aws sdk.