couchbase: allow to configure bucket replicas and default to 0.
This change allows the user to configure the number of bucket replicas
and defaults it to 0 since most of the time the user ever only needs
no replicas since it is a single container.

Note: technically this is a behavioral change since before the number
of replicas was 1 (if not defined) but this causes issues when
testing transactions (as seen in #5835).

Fixes #5835