Update toxiproxy image to ghcr.io/shopify/toxiproxy (#5996)
`shopify/toxiproxy` has moved to `ghcr.io/shopify/toxiproxy`. Currently,
our tests are using the prior and is display as part of the documentation.
In order to keep our tests and docs up-to-date with latest images
we are updating the image. Also, we are adding the proper configuration
for logging in this module.