Improve ToxiProxyContainer test and docs (#6065)
Current implementation exposed utilities to handle `ToxiProxyClient`
and proxies. However, there are some issues which demonstrate that
the container class is coupled to the client. Tests and docs have
been updated in order to promote the usage of its own `ToxiProxyClient`.
Also, a note about the number of ports reserved by Testcontainers
is added.