Remove "Starting an elasticsearch container" message in constructor
At this point in the constructor, we are not actually starting
Elasticsearch, so this message can lead to confusion. I didn't find
any of the other container classes that output a message like this
in their constructor. When the container is actually started, we
still get a log message from GenericContainer, anyway.