Enable ci for 1.18.x maintenance release
At the moment, CI runs always from `main` branch. There are some
cases when `main` branch contains changes that are not going to be
released yet and a maintenance branch should be created, so, CI
should also run.

Also, use `ref` in `actions/checkout` in order to build from the
release branch.