Apply priority execution in CI
Nowadays, around 84 jobs are executed. Core, Modules and examples
are executed in parallel which cause a bottleneck when many PRs are
raised. This commit introduces a priority to the execution

1. Core module
2. Modules
3. Examples
4. Docs examples

Although, the number of execution doesn't change at the end. It will
allow to do not take all runners at once.