gha: moby-latest: use test.docker.com for test-channel
This is just for convenience; the script at test.docker.com is identical
to the one at get.docker.com, but uses the test-channel as default.

While at it, also setting the '-f' (fail-fast), '-s' (silent, unless
failing), '-S' (force TLS), and '-L' (follow redirects) options that are
recommended when piping the output and executing it.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>