Initial import
Minimal viable product - it starts the docker container and successfully
performs a test query but does nothing beyond that.

No TLS support.

This module will require a clarification of the license - probably an
explicit carveout! - since the 'community edition' is much more
restrictive than usual. This shouldn't be a problem since the people
who actually need to use this module should already have an 'express'
or 'premium' license but we would also need a license that lets anyone
build the entire repo locally.