Update ServiceBusEmulatorContainer to not use the deprecated MSSQLServerContainer anymore.
This is a breaking change, but only in the INCUBATING Azure module. No neat way to do this without any pain.

Users of this module will need to:

change their imports from:
import org.testcontainers.containers.MSSQLServerContainer;
to:
import org.testcontainers.mssqlserver.MSSQLServerContainer;

and:

remove the <?> from the MSSQLServerContainer variable.